advanced_tensorflow/pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "advanced_flwr_tensorflow"
version = "0.1.0"
description = "Advanced Flower/TensorFlow Example"
authors = ["The Flower Authors <hello@flower.dev>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
flwr = "^1.0.0"
tensorflow-cpu = "^2.9.1"
scikit-learn = "^1.1.3"
